Added WPF client for BluetoothLE library. Modified device picker code to work from Win32 (no package identity). BluetoothClient app just connects and enumerates all services/characteristics etc - needs a UI to format the output.

+++ InTheHand.BluetoothLE/Platforms/Uap/BluetoothAdvertisingEvent.uap.cs
@@ -123,7 +123,7 @@ namespace InTheHand.Bluetooth
                     // read data
                     byte[] dataBytes = new byte[data.Data.Length - 4];
                     data.Data.CopyTo(4, dataBytes, 0, dataBytes.Length);
-                    serviceData.Add(BluetoothUuid.FromShortId(BitConverter.ToUInt32(uuidBytes, 0)), dataBytes);
+                    serviceData.Add(BluetoothUuid.FromShortId(BitConverter.ToUInt16(uuidBytes, 0)), dataBytes);
                 }
                 else if (data.DataType == BluetoothLEAdvertisementDataTypes.ServiceData16BitUuids)
                 {
